// packages/@workspace/core/src/types/ReactFlowNodeTypeNames.ts
export enum ReactFlowNodeTypeNames {
    ViewElement = "view-element",
    ViewElementPlaceholder = "view-element-placeholder",
    ViewElementBoundary = "view-element-boundary",
    ViewElementGroup = "view-element-group",
    ViewElementDeploymentNode = "view-element-deployment-node",
    ModelElement = "model-element",
    ModelElementPlaceholder = "model-element-placeholder",
    ModelElementWorkspace = "model-element-workspace"
}